rename files after class name change
