+++ cli/lib/option_parser_extended.rb
